requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
rembg[cpu]
pillow>=10.2.0
sqlalchemy>=2.0.25
aiofiles>=23.2.1
reportlab>=4.0.0
gunicorn>=21.0.0
numpy>=1.24.0
